requirements.txt
openai-whisper
sounddevice
numpy
scipy
python-dotenv
groq
pyttsx3
keyboard